+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Since last release
 * Methods to reset behavaiors of MaterialBuyPolicy (#1822)
 * More random distributions for the random number generator (#1821, #1827)
 * Added example of header code injection for facility cost (#1829)
+* Table for recording active/dormant cycle periods (#1830)
 
 **Changed:**